wip for discrete env playpen

+++ sandbox/rocky/tf/algos/npo.py
@@ -64,9 +64,9 @@ class NPO(BatchPolopt):
         else:
             valid_var = None
 
-        dist_info_vars = self.policy.dist_info_sym(obs_var, state_info_vars)
+        dist_info_vars = self.policy.dist_info_sym(tf.cast(obs_var, tf.float32), state_info_vars)
         kl = dist.kl_sym(old_dist_info_vars, dist_info_vars)
-        lr = dist.likelihood_ratio_sym(action_var, old_dist_info_vars, dist_info_vars)
+        lr = dist.likelihood_ratio_sym(tf.cast(action_var, tf.float32), old_dist_info_vars, dist_info_vars)
         if is_recurrent:
             mean_kl = tf.reduce_sum(kl * valid_var) / tf.reduce_sum(valid_var)
             surr_loss = - tf.reduce_sum(lr * advantage_var * valid_var) / tf.reduce_sum(valid_var)
